revert: split api
