+++ src/api/index.ts
@@ -1,5 +1,4 @@
 export * from './api';
-export * from './common';
 export * from './squids';
 export * from './deploy';
 export * from './secrets';
